refactor: split reset password api && splited TeamsEditView into components
