--- frontend/src/features/teams/styles/status-tags.css
@@ -1,45 +1,0 @@
-.status-tags {
-  display: flex;
-  gap: 0.35rem;
-  flex-wrap: wrap;
-}
-
-.status {
-  border-radius: 999px;
-  border: 1px solid var(--line-soft);
-  padding: 0.2rem 0.5rem;
-  font-size: 0.72rem;
-  font-weight: 700;
-  text-transform: lowercase;
-}
-
-.status--accepted {
-  border-color: rgba(16, 185, 129, 0.45);
-  background: rgba(16, 185, 129, 0.12);
-  color: #047857;
-}
-
-.status--declined {
-  border-color: rgba(239, 68, 68, 0.45);
-  background: rgba(239, 68, 68, 0.12);
-  color: #991b1b;
-}
-
-.status--invited {
-  border-color: rgba(20, 184, 166, 0.45);
-  background: rgba(20, 184, 166, 0.12);
-  color: var(--brand-700);
-}
-
-.status--pending {
-  border-color: rgba(245, 158, 11, 0.45);
-  background: rgba(245, 158, 11, 0.12);
-  color: #92400e;
-}
-
-.status--source {
-  border-color: rgba(15, 23, 42, 0.12);
-  background: rgba(15, 23, 42, 0.03);
-  color: var(--ink-700);
-  text-transform: none;
-}
