--- frontend/src/features/teams/styles/teams-edit-view.css
@@ -1,229 +1,0 @@
-.teams-edit-page {
-  gap: 1.2rem;
-}
-
-.hero-card,
-.state-card,
-.panel {
-  padding: 1.2rem;
-}
-
-.hero-card {
-  background: linear-gradient(135deg, rgba(13, 148, 136, 0.14), rgba(15, 23, 42, 0.03)), #fff;
-  border: 1px solid rgba(13, 148, 136, 0.22);
-}
-
-.hero-actions {
-  display: flex;
-  gap: 0.6rem;
-  flex-wrap: wrap;
-  margin-top: 0.8rem;
-}
-
-.action-link {
-  text-decoration: none;
-}
-
-.workspace-grid {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 1rem;
-  align-items: start;
-}
-
-.panel {
-  border: 1px solid var(--line-soft);
-}
-
-.panel-head {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.7rem;
-  margin-bottom: 0.9rem;
-}
-
-.panel-head h2 {
-  margin: 0;
-  font-family: var(--font-display);
-  font-size: 1.15rem;
-}
-
-.status-badge {
-  border-radius: 999px;
-  border: 1px solid rgba(20, 184, 166, 0.45);
-  background: rgba(20, 184, 166, 0.15);
-  color: var(--brand-700);
-  font-size: 0.74rem;
-  font-weight: 700;
-  padding: 0.2rem 0.55rem;
-}
-
-.lock-note {
-  margin-top: 0;
-}
-
-.form-grid {
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-  gap: 0.85rem;
-}
-
-.full-width {
-  grid-column: 1 / -1;
-}
-
-.form-actions {
-  display: flex;
-  gap: 0.6rem;
-  align-items: center;
-}
-
-.member-search {
-  margin-bottom: 0.75rem;
-}
-
-.member-list {
-  display: grid;
-  gap: 0.55rem;
-}
-
-.member-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 0.7rem;
-  border: 1px solid var(--line-soft);
-  border-radius: 12px;
-  background: #fff;
-  padding: 0.65rem 0.75rem;
-}
-
-.member-name,
-.member-email {
-  margin: 0;
-}
-
-.member-name {
-  font-weight: 700;
-  color: var(--ink-900);
-}
-
-.member-email {
-  font-size: 0.84rem;
-}
-
-.member-actions {
-  display: flex;
-  align-items: center;
-  gap: 0.45rem;
-}
-
-.member-note {
-  margin-top: 0.8rem;
-}
-
-.add-member-box {
-  margin-top: 0.9rem;
-  border-top: 1px solid var(--line-soft);
-  padding-top: 0.9rem;
-  display: grid;
-  gap: 0.65rem;
-}
-
-.add-member-box h3 {
-  margin: 0;
-  font-size: 1rem;
-}
-
-.btn-soft {
-  border: 1px solid var(--line-strong);
-  background: #fff;
-  border-radius: 10px;
-  padding: 0.45rem 0.7rem;
-  font: inherit;
-  font-weight: 700;
-  cursor: pointer;
-}
-
-.btn-danger {
-  border: 1px solid #dc2626;
-  background: #fee2e2;
-  color: #991b1b;
-  border-radius: 10px;
-  padding: 0.45rem 0.7rem;
-  font: inherit;
-  font-weight: 700;
-  cursor: pointer;
-}
-
-.btn-small {
-  padding: 0.28rem 0.55rem;
-  font-size: 0.8rem;
-}
-
-.btn-cancel {
-  border: 1px solid var(--line-strong);
-  background: #fff;
-  border-radius: 10px;
-  padding: 0.45rem 0.7rem;
-  font: inherit;
-  font-weight: 700;
-  cursor: pointer;
-}
-
-.modal-backdrop {
-  position: fixed;
-  inset: 0;
-  background: rgba(15, 23, 42, 0.55);
-  display: grid;
-  place-items: center;
-  z-index: 50;
-  padding: 1rem;
-}
-
-.modal-card {
-  width: min(100%, 500px);
-  background: #fff;
-  border-radius: 16px;
-  border: 1px solid var(--line-soft);
-  box-shadow: var(--shadow-lg);
-  padding: 1.2rem;
-}
-
-.modal-card h3 {
-  margin: 0;
-  font-family: var(--font-display);
-}
-
-.modal-text {
-  margin: 0.7rem 0;
-  color: var(--ink-700);
-}
-
-.modal-error {
-  margin: 0.5rem 0 0;
-}
-
-.modal-actions {
-  margin-top: 1rem;
-  display: flex;
-  justify-content: flex-end;
-  gap: 0.6rem;
-}
-
-@media (max-width: 760px) {
-  .form-grid {
-    grid-template-columns: 1fr;
-  }
-
-  .form-actions {
-    flex-direction: column;
-    align-items: stretch;
-  }
-
-  .member-row {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-}
