+++ frontend/src/services/teams/types.d.ts
@@ -11,6 +11,35 @@ export type GetTeamInfoResponse = Team & {
   can_request_to_join: boolean
 }
 
+interface CreateTeamBody {
+  name: string
+  email: string
+  organization: string
+  contact_telegram: string
+  contact_discord: string
+  is_public: boolean
+  member_ids: UserId[]
+}
+
+interface CreateTeamResponse {
+  id: number
+  name: string
+  email: string
+  captain_id: number
+  is_public: boolean
+  organization: string
+  contact_telegram: string
+  contact_discord: string
+  members: Pick<User, 'id' | 'username' | 'email' | 'full_name' | 'role'>[]
+  invitations: Invatation[]
+  join_requests: [] // TODO: add type annotation
+  my_invitation_status: boolean
+  my_join_request_status: boolean
+  is_member: boolean
+  is_captain: boolean
+  can_request_to_join: boolean
+}
+
 export type ManageJoinRequestAction = 'accept' | 'decline'
 
 export interface ResendInvatationBody {
@@ -19,4 +48,16 @@ export interface ResendInvatationBody {
 
 export interface ChangeTeamVisibilityBody {
   is_public: boolean
+}
+
+export interface UpdateTeamInfoBody {
+  name: string
+  email: string
+  organization: string
+  contact_telegram: string
+  contact_discord: string
+}
+
+export interface AddMemberBody {
+  user_id: UserId
 }
